Work around apparent cling inconsistency in application of NDEBUG. (CxxUtils-00-01-48)
	* Tagging CxxUtils-00-01-48.
	* CxxUtils/Array.icc: Work around apparent cling inconsistency in
	application of NDEBUG: use UNUSED instead of #ifndef NDEBUG.

2015-04-21  scott snyder  <snyder@bnl.gov>

	* Tagging CxxUtils-00-01-47.
	* test/pointer_list_test.cxx: Fix gcc5 compilation warning.

2015-03-16  scott snyder  <snyder@bnl.gov>

	* Tagging CxxUtils-00-01-46.
	* Root/FloatPacker.cxx (min_int): Avoid undefined behavior in bit
	shifting.
	* CxxUtils/procmaps.h, Root/procmaps.cxx: Work correctly on
	64-bit platforms.
	* test/stacktrace_test.cxx: Fix for compatibility with address
	sanitizer.

+++ Control/CxxUtils/CxxUtils/procmaps.h
@@ -18,8 +18,8 @@ class procmaps {
 public:
   struct Entry {
     Entry(const char* line);
-    unsigned int begAddress;
-    unsigned int endAddress;
+    unsigned long begAddress;
+    unsigned long endAddress;
     bool readable;
     bool writable;
     bool executable;
